[URP] MSAA Swapbuffer fix for Post-processing (v2)
fix for case 1389330 and post-processing overall. In Camera Stack + MSAA scenarios it's strange this didn't come up earlier. This being true automatically resolves the render target which leads the next camera to render to empty target thus breaking the stack. This fix checks whether it's the last camera and only does this then

FORMATTING CHANGES WERE REVERTED IN NEWEST CHANGESET