graphics/renderrequest-refactor to trunk (v3)
Jira RPFoundation: https://jira.unity3d.com/browse/GFXRPF-33
Jira Plan of Intent: https://jira.unity3d.com/browse/RND-175 

TDD: https://docs.google.com/document/d/13wTnUUrZd6u3a2x-YnAw_1at3Hdr3HlGd1jbUGZcV7Y/edit?usp=sharing

Provide a way to trigger SRP renders from script outside of the usual pipeline loop. The way this API is set-up allows us to expand functionality in the future by having a generic type.
The current issue is that API like Camera.Render offer no control to the SRP's on how these renders should be handled and break in various ways. With this API we can provide similar API to the users while giving control to the SRP's.

This PR provides the base API and implements 2 types of rendering for URP, HDRP support will be added later.

The two request types added are StandardRequest and SingleCameraRequest.
StandardRequest renders a full camera stack in URP and does not work on overlay camera's.
StandardRequest should be supported by HDRP in a future PR as well. 

SingleCameraRequest is URP specific and only renders one camera and can be executed on overlay camera's. The overlay camera will be treated as a base camera for that request.

For this API we want as many of the RequestData types to work exactly the same for both pipelines. So where it makes sense we add them to the RenderPipeline class. For those that really don't make sense we will implement them in their respective packages only.