SRP Batcher for 2D Renderers
JIRA : https://jira.unity3d.com/browse/D2D-3685

Sprite Materials are usually shared among SpriteRenderers. E.g. When you drag a new Texture to a MeshRenderer, the usual flow is to create a new Material with the material populated with the new Texture, not so in Sprite-Renderer as due to legacy reasons Texture is set as CustomProp. Also SpriteRenderer members such as Tint and Flip are also set through CustomProperties. However CustomProperties are not allowed to SRP-Batch. 2 Ways to approach this issue is to 

1) Create a new Proxy Material for every combination of the Sprite-Renderer properties (Texture, Tint, Flip etc..) 
2) Avoid using CustomProperties. This PR attempts to do this.. 

Sprite-Renderers already use an Internal PropertySheet (not Renderer CustomProps) for SecondaryTextures. And this is set/stored in Sprite Asset including (MainTex) during Build time. We utilize this to set all Textures required for the SRP Batching call when rendering Sprite-Renderer (see https://ono.unity3d.com/unity/unity/pull-request/144070/_/2d/srp/batch-batch-batch#runtime2drendererspriterendererjobscpp_n726) This solves the issue for Textures including BaseMap (MainTex) as we now populate this Internal PropertySheet with BaseMap too. For the other properties we move this to UnityPerDraw (We have moved Color to PerDraw and we can do the same for Flip if needed). Shaders are updated to reflect this change and they all work.  
Any modifications to CustomPropertySheet of Renderer will revert to non SRP Batching call as usual.

TilemapRenderer does the same as above for SRP-Batch mode. 

SpriteShapeRenderer follows pretty much the same logic but adds a custom Flattening for Material. Again instead of maintaining a new Material for each SubMesh or a new node, we create the required number of MaterialInfos for Fill and Edge Sprites and inject the required Textures in SpriteShapeRendererSRPMaterialCallback.

Particle Renderer is updated to ensure PerDraw property of RendererColor is properly set when URP is active.

Old reviews here : https://ono.unity3d.com/unity/unity/pull-request/144171/_/2d/srp/batch-batch-batch