VFX Integration in Ray Tracing (v6)
This PR enables the integration of Visual Effects into the Ray Tracing Acceleration Structure (RTAS). It also provides the necessary ray tracing shaders templates.

----------------------------------------------------------------------------------------------------------------------

JIRA Link : 
    - Epic : https://jira.unity3d.com/browse/GVFXG-138

----------------------------------------------------------------------------------------------------------------------

 The primitive used to represent particles in the RTAS is an Axis-Aligned Bounding Box (AABB), therefore, an AABB buffer is fed to the RTAS. This buffer is computed during the VFX update, either by the Update task (either for the whole system, or per output).
During the RTAS culling pass, one or more AABB buffer are retrieved from the VFXRenderer, and the corresponding masks are computed for each VFX System and/or VFX Output.

The instances are then added to the RTAS with the newly introduced method 
     void AddVFXInstances(VFXRenderer& renderer, const UInt32* vfxSystemMasks);

which use the previously introduced method 
    int AddInstance(const GraphicsBuffer& aabbBuffer, UInt32 aabbCount, UInt32 aabbOffset, bool dynamicData, const Matrix4x4& matrix, Material material, bool opaqueMaterial, const ShaderPropertySheet* properties,  UInt32 mask, UInt32 id);

----------------------------------------------------------------------------------------------------------------------

The AABB buffers are computed if and only if the platform supports Ray Tracing, and if their computation has been requested, through VFXManager::RequestRtasAabbConstruction(), which is exposed to C#, and called from HDRP at each frame when it is necessary.

The AABB buffer is created lazily the first time it is requested. Therefore, if Ray Tracing is disabled or not supported, it is never created.

There is also an option to disable all the VFXs in Ray Tracing, through VFXManager::SetRayTracingEnabled(bool enabled), and exposed as a setting in the HDRP asset.

----------------------------------------------------------------------------------------------------------------------

Limitations :

- Performance :
    - Building of Bottom-level Acceleration Structure (BLAS) : currently the BLAS is rebuilt from scratch at each frame to ensure maximal ray traversal performance. However, building the BLAS from scratch has a non-negligeable cost, and a trade-off policy can be put in place to decide when to *refit* the BLAS vs. rebuilding it from scratch.
    - Ray intersections with procedural are inherently slower than with triangles. 

- Orientation of billboard particles is approximate for shadows. This can be improved by having the light positions available in the Intersection Shader.
- Transparent objects are not considered in many ray-traced effects in HDRP. We plan to later provide a convenient optional fallback to have an opaque proxy appear in the ray-traced effects. (e.g. transparent particle is rasterized, but an opaque version of it appears in reflections).