[URP][bugfix] OnMouseOver now works with overlay cameras
This PR contains a fix for this [fogbugz case](https://fogbugz.unity3d.com/f/cases/1406991/).
This bugfix fixes the OnMouseOver function when using overlay which previously would break the functionality if not set up correctly. Also implemented upgrading for the UniversalAdditionalCameraData to make the same behavior across old and new cameras.