[VFX][UUM-540] Fix single stereo pass instanceId in SG integration
[Jira port card](https://jira.unity3d.com/browse/UUM-13742) (ISS Bug)

Fix single pass stereo rendering issue with new shader graph integration.

The setup of the instanceID for single pass stereo rendering was performed after instanceID was actually needed.
This is a simple one line fix that change the order of calls.

The fix, once found, is trivial. No need for additional testing.