[RPW][Rendering Debugger]Implementing URP Detailed Stats
https://jira.unity3d.com/browse/XPIPELINE-602

This PR does two things:
- Refactor Rendering Debugger Detailed Stats logic from HDRP into Core RP package
- Implement **"Detailed Stats" section** under Display Stats Panel on URP

The Detailed Stats panel is now implemented in RP Core using the same approach as `DebugDisplaySettingsVolume` which was moved from HDRP to Core earlier. The goal is for HDRP and URP to be able to share most of the debug panel logic, while allowing each to customize exactly what to put on their debug panels.

URP (**new**):
![image](https://media.github.cds.internal.unity3d.com/user/3380/files/653fb346-c4bc-45f1-8768-3d4a48099e47)

URP, runtime UI (**new**):
![image](https://media.github.cds.internal.unity3d.com/user/3380/files/c33d1b77-352b-4e85-8f07-19e83643be51)


HDRP (we had this before, nothing new here):
![image](https://media.github.cds.internal.unity3d.com/user/3380/files/78a75d42-0b30-469b-aaa1-29161cdd0316)