[VFX][Fix UUM-12848] Change Graph values StructuredBuffer to a ByteAddressBuffer
There is a limitation on the size of the struct used in StructuredBuffers (2048 bytes), and some large effects need a GraphValue struct that exceeds that limit. 
This PR changes the `graphValuesBuffer` into a ByteAddressBuffer to avoid that limitation.

Instead of loading the struct directly from a StructuredBuffer, we generate code that loads each member individually.