UUM-5626 terrain layer shiny
[Desc of feature/change. Links to screenshots, design docs, user docs, etc. Link to Jira if applicable. Remember reviewers may be outside your team, and not know your feature/area that should be explained more.]

For terrain layers, when the diffuse texture has an alpha channel, it is used to control smoothness. However, this is not always what the user wants. This PR adds control for the user to choose the smoothness source, either set directly via slider, or from diffuse alpha channel.

**Before:** Even though smoothness is set to 0, since diffuse is a 4 channel texture, smoothness is being controlled by alpha channel. So user has no way to control via inspector.

![image](https://media.github.cds.internal.unity3d.com/user/1582/files/36b8480e-7926-43db-b00a-58afcd055412)

**After:** Now user is able to choose whether to source smoothness from diffuse or to set directly.

![image](https://media.github.cds.internal.unity3d.com/user/1582/files/04785a05-c942-4023-9537-9349addbffc4)

![image](https://media.github.cds.internal.unity3d.com/user/1582/files/5c008309-86fc-4ea6-a235-badda41d107b)

EDIT:

Small edit to URP `TerrainLitShaderGUI.cs` so that it has as similar UI experience to `TerrainLayerInspector.cs`. The intention of doing this is so that users do not have to navigate to a different view in order to discover this new control over smoothness source.
![image](https://media.github.cds.internal.unity3d.com/user/1582/files/5f956caf-5224-4490-a73e-cd20321d45cd)

Small edit to HDRP `TerrainLitGUI.cs` so that it has as similar UI experience to `TerrainLayerInspector.cs`. The intention of doing this is so that users do not have to navigate to a different view in order to discover this new control over smoothness source.
![image](https://media.github.cds.internal.unity3d.com/user/1582/files/9cb255d2-7e63-4f84-bccb-6dc88902d808)

EDIT:

~~In the process, some code has been consolidated to `TerrainLayerUtility` that was duplicated in multiple places.~~