[URP] Fixed RenderGraph DX11 and DX12 PostPro playmode tests
Fixing the issues introduced by a previous PR in the RenderGraph tests (not on QV):
https://yamato-artifactviewer.prd.cds.internal.unity3d.com/91931afa-a671-4ea1-a1ea-332effad28b8%2FTest_Report_1%2FUtrTestReport1/TestReport.html

In some scenarios we can end up with situations where some post processing effects try to read and write to the CameraColor target, this is an invalid operations since you cannot read from a bound attachment. This PR adds a simple swap buffer mechanism to make sure we handle that edge case and don't break camera stacking.