UUM-19151 Fix URP Deferred not rendering alpha clipped SG materials
Fixes UUM-18063. Due to a `static readonly` being mutated, all Shader Graphs are using the same `RenderStateCollection` instance, and are just continously appending render states to it until the next assembly reload or Editor restart. One thing this caused was the AlphaToMask being output many times per pass, but also caused it to show up in the GBuffer pass, where it then cause the material flags channel to be used as alpha. Since that is often 0, it would end up with a 0-coverage mask and not actually render any pixels.