Fix GPU crash due to URP Forward+ (UUM-19682)
Fixes UUM-19682. Due to looking at the default renderer instead of the current camera renderer, per-object data needed in the Forward path would not be requested and thus become garbage. This caused the light loop to run infinitely, which is a no-no on GPUs and crashes after a couple of seconds.