[SRP] Fixing DirtTexture not appearing on Screen Space Lens Flare
This PR takes care of a feedback we got during the testing phase of the Screen Space Lens Flare. 
Current, SSLF write directly to the colorBuffer, so we don't have access to bloom dirt texture, so dirt is not showing on the Flares.

![Bloom_20221122_1115444_Current](https://media.github.cds.internal.unity3d.com/user/1764/files/6ef413b3-11d7-4e79-88a4-46eb1b0972a1)

Now, when doing the Screen Space Lens Flare pass, the result of this pass is added to the bloomTexture, so that when Bloom is combined in Uber shader, we have dirtTexture on the flares as well. 
This also helps intergrate flares a little better with Bloom since Bloom is thresholded before being sampled in Uber. 

As a bonus, this helps a little bit with performance on the GPU side (see perf section below)
Finally, since we're not writing to ColorBuffer, it won't trigger an additionnal MSAA resolve (which is costly for mobile). 

![f649ff6a66e64d356bc7acf6936ab5e8 (1)](https://media.github.cds.internal.unity3d.com/user/1764/files/b7fc1242-26c0-498f-904b-3b33614c386f)

Since the Screen Space Lens Flare are thresholded with Bloom, this will change the render (they will be less intense), so this will break all current tests.

Last but not least, in Uber, the Bloom is blended with the current color using the BloomIntensity directly, so that means, if Bloom intensity is 0, Screen Space Lens Flare won't show. Not sure if we can do something about it. 

Finally, this PR also removes some un-used parameters from UberPost (only in HDRP), an oversight from the past.